TECH Refactorize code and add unittests WIP
