+++ heroku2elk/lib/AMQPConnection.py
@@ -14,19 +14,24 @@ class AMQPConnectionSingleton:
     @gen.coroutine
     def get_channel(self, ioloop):
         if AMQPConnectionSingleton.__instance is None:
-            AMQPConnectionSingleton.__instance = yield AMQPConnectionSingleton.AMQPConnection().create_amqp_client(ioloop)
-        return AMQPConnectionSingleton.__instance
+            AMQPConnectionSingleton.__instance = AMQPConnectionSingleton.AMQPConnection()
+            yield AMQPConnectionSingleton.__instance.connect(ioloop)
+        return AMQPConnectionSingleton.__instance.get_channel()
 
+    @gen.coroutine
     def close_channel(self):
         if AMQPConnectionSingleton.__instance:
-            AMQPConnectionSingleton.__instance.close()
+            yield AMQPConnectionSingleton.__instance.disconnect()
         AMQPConnectionSingleton.__instance = None
+        return True
 
     class AMQPConnection:
         def __init__(self):
             self._connection = None
             self._channel = None
-            self.futureChannel = Future()
+            self._isStarted = Future()
+            self._channelClosed = Future()
+            self._connectionClosed = Future()
             self.logger = logging.getLogger("tornado.application")
             self.statsdClient = StatsClient(
                 MonitoringConfig.metrics_host,
@@ -42,7 +47,7 @@ class AMQPConnectionSingleton:
             yield self.declare_queue("heroku_production_queue")
             self.logger.info("pid:{} Exchange is declared:{} host:{} port:{}"
                              .format(os.getpid(), AmqpConfig.exchange, AmqpConfig.host, AmqpConfig.port))
-            self.futureChannel.set_result(self._channel)
+            self._isStarted.set_result(True)
 
         def declare_queue(self, name):
             future_result = Future()
@@ -55,11 +60,12 @@ class AMQPConnectionSingleton:
                                         exclusive=False, auto_delete=False)
             return future_result
 
-        def on_connection_close(self, connection):
+        def on_connection_closed(self, connection, reply_code, reply_text):
             self.logger.error("pid:{} AMQP is disconnected from exchange:{} host:{} port:{} connexion:{}"
                              .format(os.getpid(), AmqpConfig.exchange, AmqpConfig.host, AmqpConfig.port, connection))
             self._connection = None;
             self._channel = None;
+            self._connectionClosed.set_result(True)
 
         def on_connection_open(self, connection):
             self._connection = connection
@@ -91,6 +97,7 @@ class AMQPConnectionSingleton:
             """
             self.logger.warning('Channel was closed: (%s) %s', reply_code, reply_text)
             self._channel = None
+            self._channelClosed.set_result(True)
             self._connection.close()
 
         def on_delivery_confirmation(self, method_frame):
@@ -103,13 +110,25 @@ class AMQPConnectionSingleton:
             else:
                 self.statsdClient.incr('amqp.output_other', count=1)
 
-        def create_amqp_client(self, ioloop):
+        def connect(self, ioloop):
             self.logger.info("pid:{} AMQP connecting to: exchange:{} host:{} port: {}"
                         .format(os.getpid(), AmqpConfig.exchange, AmqpConfig.host, AmqpConfig.port))
             credentials = pika.PlainCredentials(AmqpConfig.user, AmqpConfig.password)
 
             pika.TornadoConnection(
                 pika.ConnectionParameters(host=AmqpConfig.host, port=AmqpConfig.port, credentials=credentials),
-                self.on_connection_open, on_close_callback=self.on_connection_close, custom_ioloop=ioloop)
+                self.on_connection_open, on_close_callback=self.on_connection_closed, custom_ioloop=ioloop)
 
-            return self.futureChannel;
+            return self._isStarted
+
+        @gen.coroutine
+        def disconnect(self):
+            self._channelClosed = Future()
+            self._connectionClosed = Future()
+            self._channel.close()
+            yield self._channelClosed
+            yield self._connectionClosed
+
+        def get_channel(self):
+            return self._channel
+
