+++ heroku2elk/lib/Statsd.py
@@ -5,7 +5,7 @@ from heroku2elk.config import MonitoringConfig
 class StatsClientSingleton:
     __instance = None
 
-    def get_instance(self):
+    def __new__(cls):
         if StatsClientSingleton.__instance is None:
             StatsClientSingleton.__instance = StatsClient(
                 MonitoringConfig.metrics_host,
