+++ heroku2elk/tools/fakelogGenerator.py
